+++ Changelog.md
@@ -59,6 +59,7 @@
  + Updated the textbox system. (Some of the old dialogue is broken as a result haven't gotten round to fixing it.)
  + Changed the intro, added some more plot.
  + You no longer gain stats when leveling up. Instead you get stat points that you can spend on your stats as you like. You can spend the points via the status menu.
+ + You get a message when you level up after a battle.
 # Known issues:
  - Update current song after victory
 # TODO: